requirements.txt
docx
PyPDF2
rtfparse
PyMuPDF
fastapi
uvicorn
python-multipart
pydantic
pdf2image
streamlit
dotenv
transformers
py7zr
Pillow
qdrant_client
sentence_transformers
docling
langchain
langchain_text_splitters
ollama